requirements.txt
amqp==2.1.4
billiard==3.5.0.5
celery==4.1.1
Django==1.10.2
django-braces==1.14.0
django-celery-beat==1.1.1
django-ckeditor==6.1.0
django-cors-headers==1.2.1
django-datatables-view==1.14.0
django-excel==0.0.10
django-js-asset==1.2.2
django-oauth-toolkit==0.12.0
djangorestframework==3.4.7
et-xmlfile==1.1.0
importlib-metadata==4.0.1
kombu==4.6.11
lml==0.1.0
numpy==1.19.5
oauthlib==2.0.1
openpyxl==3.0.9
pandas==1.1.5
pdfkit==0.5.0
pkg-resources==0.0.0
psycopg2==2.7.5
pycrypto==2.6.1
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
pyperclip==1.8.2
python-dateutil==2.7.3
python-qt==0.50
pytz==2021.1
requests==2.9.1
simplejson==3.16.0
six==1.16.0
texttable==1.6.3
typing-extensions==3.10.0.0
vine==1.3.0
xlrd==1.2.0
XlsxWriter==0.5.9
xlwt==1.3.0
zipp==3.4.1
